+++ test/luciq_http_client_test.dart
@@ -8,26 +8,26 @@ import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
-import 'package:instabug_flutter/instabug_flutter.dart';
-import 'package:instabug_flutter/src/generated/instabug.api.g.dart';
-import 'package:instabug_http_client/instabug_http_client.dart';
-import 'package:instabug_http_client/instabug_http_logger.dart';
+import 'package:luciq_flutter/luciq_flutter.dart';
+import 'package:luciq_flutter/src/generated/luciq.api.g.dart';
+import 'package:luciq_http_client/luciq_http_client.dart';
+import 'package:luciq_http_client/luciq_http_logger.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
-import 'instabug_http_client_test.mocks.dart';
+import 'luciq_http_client_test.mocks.dart';
 
 @GenerateMocks(<Type>[
-  InstabugHttpLogger,
-  InstabugHttpClient,
-  InstabugHostApi,
+  LuciqHttpLogger,
+  LuciqHttpClient,
+  LuciqHostApi,
 ])
 Future<void> main() async {
   TestWidgetsFlutterBinding.ensureInitialized();
-  final MockInstabugHostApi mHost = MockInstabugHostApi();
+  final MockLuciqHostApi mHost = MockLuciqHostApi();
 
   setUpAll(() {
-    Instabug.$setHostApi(mHost);
+    Luciq.$setHostApi(mHost);
     NetworkLogger.$setHostApi(mHost);
     when(mHost.isW3CFeatureFlagsEnabled()).thenAnswer((_)=> Future<Map<String,bool>>.value(<String, bool>{
       'isW3cCaughtHeaderEnabled': true,
@@ -45,111 +45,111 @@ Future<void> main() async {
       http.Response(json.encode(fakeResponse), 200);
 
   setUp(() {
-    url = Uri.parse('http://www.instabug.com');
+    url = Uri.parse('https://www.google.com');
   });
 
-  InstabugHttpClient buildClient({http.Client? mockClient}) {
-    final InstabugHttpClient instabugHttpClient =
-        InstabugHttpClient(client: mockClient ?? MockInstabugHttpClient());
-    instabugHttpClient.logger = MockInstabugHttpLogger();
+  LuciqHttpClient buildClient({http.Client? mockClient}) {
+    final LuciqHttpClient luciqHttpClient =
+        LuciqHttpClient(client: mockClient ?? MockLuciqHttpClient());
+    luciqHttpClient.logger = MockLuciqHttpLogger();
 
-    return instabugHttpClient;
+    return luciqHttpClient;
   }
 
   test('expect instabug http client GET to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    when<dynamic>(instabugHttpClient.client.get(url, headers: anyNamed('headers')))
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    when<dynamic>(luciqHttpClient.client.get(url, headers: anyNamed('headers')))
         .thenAnswer((_) async => mockedResponse);
-    final http.Response result = await instabugHttpClient.get(url);
+    final http.Response result = await luciqHttpClient.get(url);
     expect(result, isInstanceOf<http.Response>());
     expect(result, mockedResponse);
-    verify(instabugHttpClient.logger
+    verify(luciqHttpClient.logger
             .onLogger(mockedResponse, startTime: anyNamed('startTime')))
         .called(1);
   });
 
   test('expect instabug http client HEAD to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    when<dynamic>(instabugHttpClient.client.head(url, headers: anyNamed('headers')))
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    when<dynamic>(luciqHttpClient.client.head(url, headers: anyNamed('headers')))
         .thenAnswer((_) async => mockedResponse);
-    final http.Response result = await instabugHttpClient.head(url);
+    final http.Response result = await luciqHttpClient.head(url);
     expect(result, isInstanceOf<http.Response>());
     expect(result, mockedResponse);
-    verify(instabugHttpClient.logger
+    verify(luciqHttpClient.logger
             .onLogger(mockedResponse, startTime: anyNamed('startTime')))
         .called(1);
   });
 
   test('expect instabug http client DELETE to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    when<dynamic>(instabugHttpClient.client.delete(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    when<dynamic>(luciqHttpClient.client.delete(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
         .thenAnswer((_) async => mockedResponse);
-    final http.Response result = await instabugHttpClient.delete(url);
+    final http.Response result = await luciqHttpClient.delete(url);
     expect(result, isInstanceOf<http.Response>());
     expect(result, mockedResponse);
-    verify(instabugHttpClient.logger
+    verify(luciqHttpClient.logger
             .onLogger(mockedResponse, startTime: anyNamed('startTime')))
         .called(1);
   });
 
   test('expect instabug http client PATCH to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    when<dynamic>(instabugHttpClient.client.patch(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    when<dynamic>(luciqHttpClient.client.patch(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
         .thenAnswer((_) async => mockedResponse);
-    final http.Response result = await instabugHttpClient.patch(url);
+    final http.Response result = await luciqHttpClient.patch(url);
     expect(result, isInstanceOf<http.Response>());
     expect(result, mockedResponse);
-    verify(instabugHttpClient.logger
+    verify(luciqHttpClient.logger
             .onLogger(mockedResponse, startTime: anyNamed('startTime')))
         .called(1);
   });
 
   test('expect instabug http client POST to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    when<dynamic>(instabugHttpClient.client.post(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    when<dynamic>(luciqHttpClient.client.post(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
         .thenAnswer((_) async => mockedResponse);
-    final http.Response result = await instabugHttpClient.post(url);
+    final http.Response result = await luciqHttpClient.post(url);
     expect(result, isInstanceOf<http.Response>());
     expect(result, mockedResponse);
-    verify(instabugHttpClient.logger
+    verify(luciqHttpClient.logger
             .onLogger(mockedResponse, startTime: anyNamed('startTime')))
         .called(1);
   });
 
   test('expect instabug http client PUT to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    when<dynamic>(instabugHttpClient.client.put(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    when<dynamic>(luciqHttpClient.client.put(url, headers: anyNamed('headers'), body: anyNamed('body'), encoding: anyNamed('encoding')))
         .thenAnswer((_) async => mockedResponse);
-    final http.Response result = await instabugHttpClient.put(url);
+    final http.Response result = await luciqHttpClient.put(url);
     expect(result, isInstanceOf<http.Response>());
     expect(result.body, mockedResponse.body);
-    verify(instabugHttpClient.logger
+    verify(luciqHttpClient.logger
             .onLogger(mockedResponse, startTime: anyNamed('startTime')))
         .called(1);
   });
 
   test('expect instabug http client READ to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
+    final LuciqHttpClient luciqHttpClient = buildClient();
     const String response = 'Some response string';
-    when<dynamic>(instabugHttpClient.client.read(url,headers: anyNamed('headers')))
+    when<dynamic>(luciqHttpClient.client.read(url,headers: anyNamed('headers')))
         .thenAnswer((_) async => response);
 
-    final String result = await instabugHttpClient.read(url);
+    final String result = await luciqHttpClient.read(url);
     expect(result, isInstanceOf<String>());
     expect(result, response);
   });
 
   test('expect instabug http client READBYTES to return response', () async {
     final Uint8List response = Uint8List(3);
-    final InstabugHttpClient instabugHttpClient = buildClient(mockClient: MockClient((_) async => http.Response.bytes(response, 200)));
+    final LuciqHttpClient luciqHttpClient = buildClient(mockClient: MockClient((_) async => http.Response.bytes(response, 200)));
 
-    final Uint8List result = await instabugHttpClient.readBytes(url);
+    final Uint8List result = await luciqHttpClient.readBytes(url);
     expect(result, isInstanceOf<Uint8List>());
     expect(result, response);
   });
 
   test('expect instabug http client SEND to return response', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
+    final LuciqHttpClient luciqHttpClient = buildClient();
     final http.StreamedResponse response = http.StreamedResponse(
         const Stream<List<int>>.empty(), 200,
         contentLength: 0);
@@ -157,10 +157,10 @@ Future<void> main() async {
       ..headers[HttpHeaders.contentTypeHeader] =
           'application/json; charset=utf-8'
       ..headers[HttpHeaders.userAgentHeader] = 'Dart';
-    when<dynamic>(instabugHttpClient.client.send(request))
+    when<dynamic>(luciqHttpClient.client.send(request))
         .thenAnswer((_) async => response);
     final Future<http.StreamedResponse> responseFuture =
-        instabugHttpClient.send(request);
+        luciqHttpClient.send(request);
     request
       ..sink.add('{"hello": "world"}'.codeUnits)
       ..sink.close();
@@ -176,26 +176,26 @@ Future<void> main() async {
     expect(result.request, response.request);
     expect(await result.stream.bytesToString(),
         await response.stream.bytesToString());
-    final MockInstabugHttpLogger logger =
-        instabugHttpClient.logger as MockInstabugHttpLogger;
+    final MockLuciqHttpLogger logger =
+        luciqHttpClient.logger as MockLuciqHttpLogger;
     verify(logger.onLogger(any, startTime: anyNamed('startTime'))).called(1);
   });
 
   test('expect instabug http client CLOSE to be called', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    instabugHttpClient.close();
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    luciqHttpClient.close();
 
-    verify(instabugHttpClient.client.close());
+    verify(luciqHttpClient.client.close());
   });
 
   test('stress test for GET method', () async {
-    final InstabugHttpClient instabugHttpClient = buildClient();
-    when<dynamic>(instabugHttpClient.client.get(url, headers: anyNamed('headers')))
+    final LuciqHttpClient luciqHttpClient = buildClient();
+    when<dynamic>(luciqHttpClient.client.get(url, headers: anyNamed('headers')))
         .thenAnswer((_) async => mockedResponse);
     for (int i = 0; i < 10000; i++) {
-      await instabugHttpClient.get(url);
+      await luciqHttpClient.get(url);
     }
-    verify(instabugHttpClient.logger
+    verify(luciqHttpClient.logger
             .onLogger(mockedResponse, startTime: anyNamed('startTime')))
         .called(10000);
   });